--- app/src/dev/res/layout/content_dev_settings.xml
@@ -1,8 +1,0 @@
-<androidx.fragment.app.FragmentContainerView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:name="com.atomicrobot.carbon.ui.devsettings.DevSettingsFragment"
-        android:id="@+id/fragment_dev_settings"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
