--- app/src/main/java/com/atomicrobot/carbon/ui/ActivityScope.kt
@@ -1,8 +1,0 @@
-package com.atomicrobot.carbon.ui
-
-import javax.inject.Scope
-
-@Scope
-@MustBeDocumented
-@Retention(AnnotationRetention.RUNTIME)
-annotation class ActivityScope
