--- app/src/main/java/com/atomicrobot/carbon/ui/SimpleSnackbarMessage.kt
@@ -1,28 +1,0 @@
-package com.atomicrobot.carbon.ui
-
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.Observer
-
-/**
- * A SingleLiveEvent used for Snackbar messages. Like a [SingleLiveEvent] but also prevents
- * null messages and uses a custom observer.
- *
- * Note that only one observer is going to be notified of changes.
- */
-class SimpleSnackbarMessage : SingleLiveEvent<String>() {
-    fun observe(owner: LifecycleOwner, observer: SnackbarObserver) {
-        super.observe(owner, Observer { message ->
-            if (message == null) {
-                return@Observer
-            }
-            observer.onNewMessage(message)
-        })
-    }
-
-    interface SnackbarObserver {
-        /**
-         * Called when there is a new message to be shown.
-         */
-        fun onNewMessage(message: String)
-    }
-}
