--- app/src/main/java/com/atomicrobot/carbon/util/databinding/ViewAdapters.kt
@@ -1,12 +1,0 @@
-package com.atomicrobot.carbon.util.databinding
-
-import android.view.View
-import androidx.databinding.BindingAdapter
-
-object ViewAdapters {
-    @JvmStatic
-    @BindingAdapter("visible")
-    fun setVisible(view: View, visible: Boolean) {
-        view.visibility = if (visible) View.VISIBLE else View.INVISIBLE
-    }
-}
