--- app/src/main/java/com/atomicrobot/carbon/util/recyclerview/ArrayAdapter.kt
@@ -1,19 +1,0 @@
-package com.atomicrobot.carbon.util.recyclerview
-
-import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.ViewHolder
-import java.util.*
-
-abstract class ArrayAdapter<Item, ItemViewHolder : ViewHolder> : RecyclerView.Adapter<ItemViewHolder>() {
-
-    private var items: List<Item> = ArrayList()
-
-    fun setItems(items: List<Item>) {
-        this.items = items
-        notifyDataSetChanged()
-    }
-
-    override fun getItemCount() = items.size
-
-    protected fun getItemAtPosition(position: Int) = items[position]
-}
